requirements.txt
PyQt6>=6.4.0
yt-dlp>=2024.1.0
PyYAML>=6.0
imageio-ffmpeg>=0.5.0
